feat: sentry error handler improvements
