+++ api/src/index.ts
@@ -1,11 +1,46 @@
+import type express from 'express';
+import * as Sentry from "@sentry/node";
 import { createApp } from './app';
-import { startServer } from './graphql/server';
+import { addGraphqlRoutes } from './graphql/server';
 import { getChildLogger } from './logger';
 import { closeRedis } from './redis';
+import { Server, createServer as createHttpServer } from 'http';
+import { PORT } from './env';
+import { errorHandler } from './middleware/errors';
+import { initRedis } from './redis';
+import "express-async-errors";
 
 const logger = getChildLogger({ msgPrefix: 'SERVER' });
 
 const { app, router } = createApp();
+
+async function startServer(
+  app: express.Application,
+  router: express.Router,
+): Promise<Server> {
+  await addGraphqlRoutes(router);
+
+  logger.debug('Starting server');
+
+  if (process.env.REDIS_TLS_URL) {
+    await initRedis();
+  }
+
+  const basePath = process.env.BASE_PATH || '/';
+  app.use(basePath, router);
+
+  if (process.env.SENTRY_DSN) {
+    app.use(Sentry.Handlers.errorHandler());
+  }
+
+  app.use(errorHandler);
+
+  const httpServer = createHttpServer(app);
+  return httpServer.listen({ port: PORT }, () => {
+    logger.info(`Running on ${PORT}`);
+  });
+}
+
 
 /* istanbul ignore next @preserve */
 startServer(app, router)
