+++ api/src/middleware/errors.ts
@@ -1,14 +1,11 @@
 import { NextFunction, Request, Response } from "express";
 import { getChildLogger } from "../logger";
-import { CustomError } from "../errors/CustomError";
 
 const logger = getChildLogger({ msgPrefix: 'Generic Error Handler', });
 
 export const errorHandler = (err: Error, req: Request, res: Response, next: NextFunction) => {
+  const errorId = (res as { sentry?: string }).sentry;
   err.cause = err.cause || 'Unknown';
-  logger.error({
-    msg: err.message,
-  });
-
-  res.status(500).send({ errors: err.message });
+  logger.error(err, err.message, { errorId });
+  res.status(500).send({ errors: err.message, errorId });
 };
